+++ src/server.js
@@ -27,6 +27,9 @@ if (cluster.isMaster && (isDev || isProd)) {
   var swagger = require("swagger-node-express");
   var autoinc = require('mongoose-id-autoinc');
 
+  // Setup translations
+  var i18n = require('./i18n');
+
   var middleware = require('./middleware');
 
   var TWO_WEEKS = 1000 * 60 * 60 * 24 * 14;
